+++ arpg-timeline/src/middleware.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 
 const validRoutes = ["/", "/looking-for-moderators", "/privacy", "/api/revalidate", "/api/webhook"];
-const routePrefixes = ["/docs/html/", "/docs/obs/", "/embed/season-widget/"];
+const routePrefixes = ["/docs/html/", "/docs/obs/", "/embed/season-widget/", "/dashboard/"];
 
 function isValidRoute(pathname: string): boolean {
     if (validRoutes.includes(pathname)) {
